Clean up tests and reduce duplication
And test the validity of the generated XML in the TwoWaySpec.
